refactor: Refactor Iroha Javascript library
